Remove todos.json history and see if it breaks everything
I see the todos.json file is used in the script to do before and after comparisons between runs but I think that should still work since the filter-history will only remove the historical versions of the file but keep the current version?

If not and everything goes awry I will just return it.