Handle no cache files existing
I managed to wipe out the history of files I shouldn't have. This will reset change tracking of repositories and followers, which sucks, but it is not that much of a biggie. It will get recreated with time for new entries.

I made a change for the script to be able to run from the ground zero because I think I will lose these cached values in the future again.

I also made a note to not rely on cache files at all and instead use workflow artifacts for this as they seem like they would work better.