Generate and push repository lists ordered by various criteria
I had prototyped this in a private repository and I think it can go in here now. I have a few improvements down in line I want to do still which I am tracking in the original repository for now but will turn into to-do items or code lines here soon.