Drop issuesAndPrs as a JSON file artifact for debugging
I don't want this to clog up the logs in case there is a problem with it or worse, in case there isn't. But I need to see the contents of the response in case there is a problem with it. I will drop it as an artifact and upload it on script failure. I will add the artifact upload step in the next commit.