Upload debugging artifacts in case of failure
The search endpoint has a secondary rate limit and sometimes experiences exhaustion even if the primary API rate limit is safe. In that case I made the script drop the search API response JSON as a file to upload as an artifact for debugging which is what this script will do now.